+++ src/anomalib/data/image/__init__.py
@@ -9,6 +9,7 @@ This module contains the supported image datasets for Anomalib.
 from enum import Enum
 
 from .btech import BTech
+from .dataframe import Dataframe
 from .datumaro import Datumaro
 from .folder import Folder
 from .kolektor import Kolektor
@@ -20,6 +21,7 @@ class ImageDataFormat(str, Enum):
     """Supported Image Dataset Types."""
 
     BTECH = "btech"
+    DATAFRAME = "dataframe"
     DATUMARO = "datumaro"
     FOLDER = "folder"
     FOLDER_3D = "folder_3d"
@@ -29,4 +31,4 @@ class ImageDataFormat(str, Enum):
     VISA = "visa"
 
 
-__all__ = ["BTech", "Datumaro", "Folder", "Kolektor", "MVTec", "Visa"]
+__all__ = ["BTech", "Dataframe", "Datumaro", "Folder", "Kolektor", "MVTec", "Visa"]