+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- 🚀 Add Dataframe datamodule by @manuelkonrad in https://github.com/openvinotoolkit/anomalib/pull/2403
+
 ### Changed
 
 ### Deprecated